+++ API/Collection/ContentTypeGroupCollection.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Kaliop\eZMigrationBundle\API\Collection;
+
+/**
+ * @todo add phpdoc to suggest typehinting
+ */
+class ContentTypeGroupCollection extends AbstractCollection
+{
+    protected $allowedClass = 'eZ\Publish\API\Repository\Values\ContentType\ContentTypeGroup';
+}